refactor(app/cms): :lipstick: update form UI
