+++ apps/cms/package.json
@@ -2,6 +2,7 @@
   "name": "@meme-dock/cms",
   "version": "0.0.1",
   "private": true,
+  "type": "module",
   "dependencies": {
     "@hookform/resolvers": "^5.0.1",
     "@radix-ui/react-dialog": "^1.1.14",
