+++ package.json
@@ -1,55 +1,55 @@
 {
-    "name": "@meme-dock/source",
-    "version": "0.0.0",
-    "license": "MIT",
-    "scripts": {},
-    "private": true,
-    "devDependencies": {
-        "@eslint/compat": "^1.1.1",
-        "@eslint/eslintrc": "^2.1.1",
-        "@eslint/js": "^9.8.0",
-        "@next/eslint-plugin-next": "^15.2.4",
-        "@nx/eslint": "21.1.2",
-        "@nx/eslint-plugin": "21.1.2",
-        "@nx/js": "21.1.2",
-        "@nx/next": "21.1.2",
-        "@nx/react": "^21.1.2",
-        "@nx/vite": "21.1.2",
-        "@nx/web": "21.1.2",
-        "@swc-node/register": "~1.9.1",
-        "@swc/cli": "~0.6.0",
-        "@swc/core": "~1.5.7",
-        "@swc/helpers": "~0.5.11",
-        "@tailwindcss/postcss": "^4.1.8",
-        "@types/node": "^20.0.0",
-        "@types/react": "19.0.0",
-        "@types/react-dom": "19.0.0",
-        "@vitejs/plugin-react": "^4.2.0",
-        "@vitejs/plugin-react-swc": "^3.5.0",
-        "@vitest/ui": "^3.0.0",
-        "eslint": "^9.8.0",
-        "eslint-config-next": "^15.2.4",
-        "eslint-config-prettier": "^10.0.0",
-        "eslint-plugin-import": "2.31.0",
-        "eslint-plugin-jsx-a11y": "6.10.1",
-        "eslint-plugin-react": "7.35.0",
-        "eslint-plugin-react-hooks": "5.0.0",
-        "jiti": "2.4.2",
-        "jsdom": "~22.1.0",
-        "nx": "21.1.2",
-        "postcss": "^8.5.4",
-        "prettier": "^2.6.2",
-        "tailwindcss": "^4.1.8",
-        "tslib": "^2.3.0",
-        "typescript": "~5.7.2",
-        "typescript-eslint": "^8.19.0",
-        "vite": "^6.0.0",
-        "vite-plugin-dts": "~4.5.0",
-        "vitest": "^3.0.0"
-    },
-    "dependencies": {
-        "next": "~15.2.4",
-        "react": "19.0.0",
-        "react-dom": "19.0.0"
-    }
+  "name": "@meme-dock/source",
+  "version": "0.0.0",
+  "license": "MIT",
+  "scripts": {},
+  "private": true,
+  "devDependencies": {
+    "@eslint/compat": "^1.1.1",
+    "@eslint/eslintrc": "^2.1.1",
+    "@eslint/js": "^9.8.0",
+    "@next/eslint-plugin-next": "^15.2.4",
+    "@nx/eslint": "21.1.2",
+    "@nx/eslint-plugin": "21.1.2",
+    "@nx/js": "21.1.2",
+    "@nx/next": "21.1.2",
+    "@nx/react": "^21.1.2",
+    "@nx/vite": "21.1.2",
+    "@nx/web": "21.1.2",
+    "@swc-node/register": "~1.9.1",
+    "@swc/cli": "~0.6.0",
+    "@swc/core": "~1.5.7",
+    "@swc/helpers": "~0.5.11",
+    "@tailwindcss/postcss": "^4.1.8",
+    "@types/node": "^20.0.0",
+    "@types/react": "19.0.0",
+    "@types/react-dom": "19.0.0",
+    "@vitejs/plugin-react": "^4.2.0",
+    "@vitejs/plugin-react-swc": "^3.5.0",
+    "@vitest/ui": "^3.0.0",
+    "eslint": "^9.8.0",
+    "eslint-config-next": "^15.2.4",
+    "eslint-config-prettier": "^10.0.0",
+    "eslint-plugin-import": "2.31.0",
+    "eslint-plugin-jsx-a11y": "6.10.1",
+    "eslint-plugin-react": "7.35.0",
+    "eslint-plugin-react-hooks": "5.0.0",
+    "jiti": "2.4.2",
+    "jsdom": "~22.1.0",
+    "nx": "21.1.2",
+    "postcss": "^8.5.4",
+    "prettier": "^2.6.2",
+    "tailwindcss": "^4.1.8",
+    "tslib": "^2.3.0",
+    "typescript": "~5.7.2",
+    "typescript-eslint": "^8.19.0",
+    "vite": "^6.0.0",
+    "vite-plugin-dts": "~4.5.0",
+    "vitest": "^3.0.0"
+  },
+  "dependencies": {
+    "next": "~15.2.4",
+    "react": "19.0.0",
+    "react-dom": "19.0.0"
+  }
 }
